waypoint and mission manager scaffolding
