+++ cougars_nav/src/mission_manager.cpp
@@ -1,0 +1,14 @@
+
+#include "rclcpp/rclcpp.hpp"
+
+class MissionManager : public rclcpp::Node
+{
+}
+
+int main(int argc, char * argv[])
+{
+    rclcpp::init(argc, argv);
+    rclcpp::spin(std::make_shared<MissionManager>()); 
+    rclcpp::shutdown();
+    return 0;
+}
