+++ cougars_nav/src/waypoint_manager.cpp
@@ -1,0 +1,14 @@
+
+#include "rclcpp/rclcpp.hpp"
+
+class WaypointManager : public rclcpp::Node
+{
+}
+
+int main(int argc, char * argv[])
+{
+    rclcpp::init(argc, argv);
+    rclcpp::spin(std::make_shared<WaypointManager>()); 
+    rclcpp::shutdown();
+    return 0;
+}
